psr fix

+++ application/modules/faq/controllers/admin/Index.php
@@ -105,7 +105,7 @@ class Index extends \Ilch\Controller\Admin
             );
 
             $validation = Validation::create($this->getRequest()->getPost(), [
-                'catId' => 'required|numeric|integer|min:1',
+                'catId' => 'required|numeric|integer|min:1|exists:' . $categoryMapper->tablename,
                 'question' => 'required',
                 'answer' => 'required'
             ]);
@@ -132,7 +132,7 @@ class Index extends \Ilch\Controller\Admin
 
     public function delFaqAction()
     {
-        if ($this->getRequest()->isSecure()) {
+        if ($this->getRequest()->isSecure() && !empty($this->getRequest()->getParam('id'))) {
             $faqMapper = new FaqMapper();
             $faqMapper->delete($this->getRequest()->getParam('id'));
